refactor(frontend): make nullspace replace atomic and split module
Address CodeRabbit review on the nullspace generation PR.

Replace the iterative remove/push field mutation with a single
field.setValue assignment. The calibration form auto-submits on every
value change, so clearing and re-adding rows one-by-one could emit
intermediate empty/partial nullspaceVectors states and persist them if
a submit landed mid-sequence. One atomic assignment emits a single
change.

Split nullspaceComputation.ts (was at the 300-line max-lines limit)
into focused modules under nullspace/: constants, rref (RREF + basis
extraction), sparsity (Gram-Schmidt + L1 search), and postProcessing
(normalize/round/threshold/reconstruct). The orchestrator keeps the
public computeNullspaceFromAllocation API.
