+++ src/features/settings/forms/calibration/nullspace/constants.ts
@@ -1,0 +1,6 @@
+export const COMPUTATION_EPSILON = 1e-9;
+export const DEADZONE_THRESHOLD_VALUE = 0.2;
+export const ROUND_PRECISION = 100;
+export const NUM_THRUSTERS = 8;
+export const NUM_MOTION_DOFS = 6;
+export const L1_SCAN_STEPS = 2000;
